feature: add image scaling option to settings
https://github.com/Attacktive/Wallhavend/issues/7
